Create Co-Ordinates.py
HACKERRANK-PYTHON-LIST COMPREHENSION

You are given three integers x,y,z  representing the dimensions of a cuboid along with an integer n . Print a list of all possible coordinates given by (i,j,k) on a 3D grid where the sum of  is not equal to n. Here , 0 <= i,j,k <= x,y,z respectively. Print an array of the elements that do not sum to n. Print the list in lexicographic increasing order.