Create SubSring.py
HACKERRANK-COUNT-SUBSTRING

The user enters a string and a substring. You have to print the number of times that the substring occurs in the given string. String traversal will take place from left to right, not from right to left.
NOTE: String letters are case-sensitive.

For eg.  If string="ABCDCDC" and sub_string="CDC"
then the output is 
2

note:count() will give 1 as output which is wrong
since 2 occurences of "CDC" are overlapped